refactor auth controllers
