+++ server/controllers/profileController.js
@@ -1,0 +1,155 @@
+const User = require("../models/User");
+const jwt = require("jsonwebtoken");
+const crypto = require("crypto");
+const sendEmail = require("../utils/sendEmail");
+const { getPasswordResetTemplate } = require("../utils/emailTemplates");
+
+// Get user profile
+exports.getProfile = async (req, res, next) => {
+  try {
+    const user = await User.findById(req.user.id).select("-password");
+    res.json(user);
+  } catch (err) {
+    next(err);
+  }
+};
+
+// Update user profile
+exports.updateProfile = async (req, res, next) => {
+  try {
+    const { name, email } = req.body;
+    const updateFields = {};
+    if (name) updateFields.name = name;
+    if (email) updateFields.email = email;
+
+    const user = await User.findByIdAndUpdate(
+      req.user.id,
+      { $set: updateFields },
+      { new: true },
+    ).select("-password");
+
+    res.json(user);
+  } catch (err) {
+    next(err);
+  }
+};
+
+// Change password
+exports.changePassword = async (req, res, next) => {
+  try {
+    const { currentPassword, newPassword } = req.body;
+    const user = await User.findById(req.user.id);
+    const isMatch = await user.comparePassword(currentPassword);
+    if (!isMatch) {
+      return res.status(400).json({ msg: "Current password is incorrect" });
+    }
+
+    user.password = newPassword;
+    await user.save();
+
+    res.json({ msg: "Password updated successfully" });
+  } catch (err) {
+    next(err);
+  }
+};
+
+// Forgot Password
+exports.forgotPassword = async (req, res, next) => {
+  try {
+    const user = await User.findOne({ email: req.body.email });
+
+    if (!user) {
+      return res.status(400).json({ msg: "There is no user with that email" });
+    }
+
+    // Get reset token
+    const resetToken = user.getResetPasswordToken();
+
+    await user.save({ validateBeforeSave: false });
+
+    // Create reset url
+    // Assumes frontend is running on localhost:3000 during dev or the deployed URL
+    const frontendUrl = process.env.FRONTEND_URL || "http://localhost:3000";
+    const resetUrl = `${frontendUrl}/reset-password/${resetToken}`;
+
+    // Only print to console during local development for easy testing
+    if (process.env.NODE_ENV === "development") {
+      console.log("\n=======================================================");
+      console.log("ðŸš€ DEV MODE: PASSWORD RESET LINK GENERATED");
+      console.log(resetUrl);
+      console.log("=======================================================\n");
+    }
+
+    const message = `You are receiving this email because you (or someone else) has requested the reset of a password. Please make a PUT request to: \n\n ${resetUrl}`;
+
+    try {
+      // We still try to send the email, but if it takes too long, they already have the link above!
+      await sendEmail({
+        email: user.email,
+        subject: "Password reset token",
+        message,
+        html: getPasswordResetTemplate(user.name, resetUrl),
+      });
+
+      res.status(200).json({ success: true, data: "Email sent successfully" });
+    } catch (err) {
+      console.error("Email sending failed:", err);
+
+      // Reset the token fields since the email failed
+      user.resetPasswordToken = undefined;
+      user.resetPasswordExpire = undefined;
+      await user.save({ validateBeforeSave: false });
+
+      // Return a proper 500 error in production
+      return res
+        .status(500)
+        .json({ msg: "Email could not be sent. Please try again later." });
+    }
+  } catch (err) {
+    next(err);
+  }
+};
+
+// Reset Password
+exports.resetPassword = async (req, res, next) => {
+  try {
+    // Get hashed token
+    const resetPasswordToken = crypto
+      .createHash("sha256")
+      .update(req.params.token)
+      .digest("hex");
+
+    const user = await User.findOne({
+      resetPasswordToken,
+      resetPasswordExpire: { $gt: Date.now() },
+    });
+
+    if (!user) {
+      return res.status(400).json({ msg: "Invalid token" });
+    }
+
+    // Set new password
+    user.password = req.body.password;
+    user.resetPasswordToken = undefined;
+    user.resetPasswordExpire = undefined;
+    await user.save();
+
+    // Create JWT token and log user in automatically (optional)
+    const payload = { user: { id: user.id } };
+    jwt.sign(
+      payload,
+      process.env.JWT_SECRET,
+      { expiresIn: "5d" },
+      (err, token) => {
+        if (err) throw err;
+        res.json({
+          msg: "Password reset successful",
+          token,
+          user: { id: user.id, name: user.name, email: user.email },
+        });
+      },
+    );
+  } catch (err) {
+    next(err);
+  }
+};
